Add check for 2000 char limit

+++ commands/run.js
@@ -36,6 +36,11 @@ module.exports = {
 	
 		  if (typeof evaled !== "string")
 			evaled = evaled.toString();
+			
+			if (evaled.length >= 2000) {
+				message.reply(`Output was longer than 2000 characters (${evaled.length} to be exact!) You can find it in the console.`);
+				return console.log(evaled);
+			}
 	
 			message.channel.send(`**Evaluation successful.**\n\n:inbox_tray: Input:\n\`\`\`sh\n${code}\n\`\`\`\n\n:outbox_tray: Output:\n\`\`\`sh\n${clear(evaled)}\n\`\`\`\n\`Evaluation Completed\``);
 			msg.delete();
